+++ packages/availity-workflow-settings/workflow.js
@@ -23,7 +23,9 @@ const development = {
   hot: true,
 
   // Enable or disable react-hot-loader
-  hotLoader: true
+  hotLoader: true,
+
+  hotLoaderEntry: /\/App\.jsx?/
 
   // Allows developers to override the babel-preset-env target to match their developer environment.  This is benefecial if
   // a developer is doing their primary development environment in a browser like Chrome 57+ that already supports a lot